refactor: naming 변경, kakao 디렉토리 생성
