+++ src/main/kotlin/mara/server/auth/kakao/KaKaoUserLogout.kt
@@ -1,4 +1,4 @@
-package mara.server.auth
+package mara.server.auth.kakao
 
 data class KaKaoUserLogout(
     val id: Long,
