+++ src/main/kotlin/mara/server/auth/kakao/KakaoApiClient.kt
@@ -1,4 +1,4 @@
-package mara.server.auth
+package mara.server.auth.kakao
 
 import mara.server.util.logger
 import org.springframework.beans.factory.annotation.Value
@@ -34,18 +34,17 @@ class KakaoApiClient(
     fun getRedirectUri(status: String): String {
         val os = System.getProperty("os.name")
         log.info("OS : {}", os)
-//        if (os.contains("Mac") || os.contains("Windows")) return "http://localhost:8080/member/login"
         if (status == "local")return ""
         if (status == "prod") return ""
         return "http://localhost:8080/users/kakao-login"
     }
 
-    fun requestAccessToken(code: String, status: String): String {
+    fun requestAccessToken(authorizedCode: String, status: String): String {
         val url = "$authUrl/oauth/token"
         val httpHeaders = HttpHeaders()
         httpHeaders.contentType = MediaType.APPLICATION_FORM_URLENCODED
         val body: MultiValueMap<String, String> = LinkedMultiValueMap()
-        body.add("code", code)
+        body.add("code", authorizedCode)
         body.add("grant_type", "authorization_code")
         body.add("client_id", clientId)
         body.add("client_secret", secret)
